various modifications in preparation for initial release
