+++ .rubocop.yml
@@ -16,3 +16,23 @@ Lint/ConstantDefinitionInBlock:
 Layout/LineLength:
   Exclude:
     - 'spec/**/*'
+
+RSpec/FilePath:
+  SpecSuffixOnly: true
+
+RSpec/MultipleExpectations:
+  Max: 4
+
+RSpec/ExampleLength:
+  Max: 10
+  Exclude:
+    - spec/examples/*
+
+RSpec/DescribeClass:
+  Exclude:
+    - spec/examples/*
+    - spec/features/*
+
+RSpec/MethodLength:
+  Exclude:
+    - spec/examples/*
